py/pyproject.toml
[project]
name = "py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "plotly-express>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
